% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanBedopsOutput.R
\name{CleanBedopsOutput}
\alias{CleanBedopsOutput}
\title{Clean bedops output}
\usage{
CleanBedopsOutput(filestub, segfile_name)
}
\arguments{
\item{filestub}{= dir (with trailling slash) containing *_bedops
segfile_name = label of cohort, eg 'TGCT'}
}
\value{
Clean version of bedops file
}
\description{
Cleans output from bedops which leaves segments spanning whole chrs which screw up downstream steps
}
