% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNAsDriversComplexHeatmap.R
\name{CNAsDriversComplexHeatmap}
\alias{CNAsDriversComplexHeatmap}
\title{Heatmap to show CNA status in a set of driver genes across a cohort of tumour samples
Dendrogram automatically goes alongside
Bars classifying samples in lots of ways produced alongside heatmap}
\usage{
CNAsDriversComplexHeatmap(segfile_dir, segfile_name, bed_file, driver_file)
}
\arguments{
\item{segfile_dir}{= dir with all *_subclones.txt files are
segfile_name = label of cohort, eg 'TGCT'
bed_file = tab sep file, 7 cols, chr,start,end,strand,ens,ens,genename
driver_file = tab sep file, colnames SYMBOL (of gene) and TIER (1,2,etc) existing}
}
\value{
Heatmap, dendrogram of CNAs v drivers, with classifying bars
}
\description{
Create heatmap showing CNA status at driver genes using complexheatmap
homdel, loh, otherloss, nochange, gain, biggain
}
