% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeSegments_test.R
\name{CodeSegments_test}
\alias{CodeSegments_test}
\title{Code all segments in all samples in the cohort}
\usage{
CodeSegments_test(segfile_dir, segfile_name)
}
\arguments{
\item{segfile_dir}{= where segfile will go
segfile_name = cohort name}
}
\value{
A table for use with bedtools etc
}
\description{
Codes all segments as a particular CNA type from across the cohort
}
