% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateDPinput_standard_vcf.R
\name{GenerateDPinput_mutect_vcf}
\alias{GenerateDPinput_mutect_vcf}
\title{Generate DP input}
\usage{
GenerateDPinput_mutect_vcf(
  tumourplatekey,
  normalplatekey,
  gender,
  vcffilepath,
  rhoandpsifilepath,
  subcloneshg38filepath,
  output_loci_file,
  output_file,
  output_DPinput_file
)
}
\arguments{
\item{Loci}{of SNP positions used in CNV call
Allele frequencies of same SNP positions
File containing purity and ploidy of tumour sample
Subclones file from Battenberg containing copy number segments across genome
Sex of patient}
}
\value{
A file containing all snv loci,
their mutation copy number status,
and their CCF (cancer cell fraction)
}
\description{
Converts snv counts from vcf into input format for DPClust
}
