% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomewideStackedBar.R
\name{GenomewideStackedBar}
\alias{GenomewideStackedBar}
\title{Prep data for genome wide stacked barplot
Create stacked bar plot}
\usage{
GenomewideStackedBar(filestub, segfile_name)
}
\value{
Prepped data
Stacked barplot
}
\description{
Take output from bedtools intersect and prepare data for stacked barplot
Create stacked barplot for 6 different CNA types
homdel, loh, otherloss, nochange, gain, biggain
}
