% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomewideStackedBarPlot.R
\name{GenomewideStackedBarPlot}
\alias{GenomewideStackedBarPlot}
\title{Create stacked bar plot}
\usage{
GenomewideStackedBarPlot(filestub, segfile_name, number_samples)
}
\arguments{
\item{filestub}{= dir (with trailling slash) containing *_partitioned_regions_all_CNA_types_overlapped.out files for all cna types
segfile_name = label of cohort, eg 'TGCT'}
}
\value{
Prepped data
Stacked barplot
}
\description{
Create stacked barplot for 6 different CNA types
homdel, loh, otherloss, nochange, gain, biggain, split into loss, no change, and gain
}
