% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeWideStackedBarPlotGains.R
\name{GenomewideStackedBarPlotGains}
\alias{GenomewideStackedBarPlotGains}
\title{Create stacked bar plot}
\usage{
GenomewideStackedBarPlotGains(filestub, segfile_name, number_samples)
}
\arguments{
\item{filestub}{= dir (with trailling slash) containing *_partitioned_regions_all_CNA_types_overlapped.out files for all cna types
segfile_name = label of cohort, eg 'TGCT'}
}
\value{
Prepped data
Stacked barplot
}
\description{
Create stacked barplot for gains
gain, biggain
}
