% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelGenomecov.R
\name{LabelGenomecov}
\alias{LabelGenomecov}
\title{Collate all segments in all samples in the cohort}
\usage{
LabelGenomecov(filestub, segfile_name)
}
\arguments{
\item{filestub}{= dir (with trailling slash) containing *regions_normalised_out files for all cna types
segfile_name = label of cohort, eg 'TGCT'}
}
\value{
writes same name file with cna types labelled
}
\description{
Label bedtools_genomecov files with type of CNA
}
