% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepDataForSummary.R
\name{PrepDataForSummary}
\alias{PrepDataForSummary}
\alias{prepDataForSummary}
\title{Make summary table with filepaths for required files (subclones and purity)}
\usage{
prepDataForSummary(qctabledir, subclonesdir)
}
\arguments{
\item{Filepath}{of QualityControl.tsv file
Dir of subclones and cellularity files}
}
\value{
A summary table for producing CNA summary plots
}
\description{
Converts snv counts from vcf into input format for DPClust
}
