% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDPClust.R, R/RunDPClust_Beagle_NF.R
\name{RunDPClust_Beagle_nf}
\alias{RunDPClust_Beagle_nf}
\alias{RunDPClust}
\title{Runs DPClust using input provided in GenerateDPinputt2BamAC_Beagle_nf}
\usage{
RunDPClust(
  sampledir,
  participantid,
  tumourplatekey,
  normalplatekey,
  dpinfofilepath,
  gender,
  vcffilepath,
  rhoandpsifilepath,
  subcloneshg38filepath,
  run
)

RunDPClust_Beagle_nf(
  tumourplatekey,
  normalplatekey,
  rhoandpsifilepath,
  dpinfofilepath,
  output_folder
)
}
\arguments{
\item{Tumour}{platekey
Germline platekey
Filepath to purity and ploidy file from Battenberg
Filepath to input file generated by GenerateDPinputt2BamAC_Beagle_nf
Output directory}
}
\value{
Cluster assignment for all SNVs
Plots showing the clusters
OptimaInfo file summarising the clusters
And the CCF and number of mutations assigned to each

Cluster assignment for all SNVs
Plots showing the clusters
OptimaInfo file summarising the clusters
And the CCF and number of mutations assigned to each
}
\description{
Clusters mutations by CCF using Dirichlet Process

Clusters mutations by CCF using Dirichlet Process
}
