% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepDataForSummary.R
\docType{data}
\name{PrepDataForSummary}
\alias{PrepDataForSummary}
\alias{x}
\title{Make summary table with filepaths for required files (subclones and purity)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
x
}
\arguments{
\item{Filepath}{of QualityControl.tsv file
Dir of subclones and cellularity files}
}
\value{
A summary table
}
\description{
Converts snv counts from vcf into input format for DPClust
}
\keyword{datasets}
